Optimize 2017 day 2 solution
I noticed, that it also makes sense to have the lines of the input as
sorted int-arrays for the first part for today. Then I just have to pick
the first and the last element for calculating the checksum.
This also brings a signature change for the method used to solve part 2.